Wire up and fix example programs
The 5 ack/router example files lived at the workspace-root examples/ directory
where Cargo never compiled them (the kincir crate's examples belong in
kincir/examples/), so they had bit-rotted - including invalid println! calls
that never compiled. The mqtt-to-rabbitmq-example crate was also missing from
the workspace members and pointed at the wrong kincir path.

- Move ack_example, router_ack_example, kafka_ack_example, mqtt_ack_example and
  rabbitmq_ack_example into kincir/examples/ so they are real
  'cargo run --example <name>' targets, and repair them:
  - fix invalid println!("\n" + ...) / println!("=".repeat(n)) calls
  - use the .boxed() adapter for AckRouter publishers and annotate handler
    closures as HandlerFunc (router_ack_example)
  - import AckHandle so handle accessors resolve; make the subscriber mut;
    use StatsSnapshot fields instead of non-existent methods; avoid borrowing
    a dropped temporary in unwrap_or (ack_example)
- Add the orphaned mqtt-to-rabbitmq-example crate to the workspace members and
  point its dependency at ../../kincir
- Drop an unused Logger import in kafka-example

All examples now compile; cargo clippy --all-targets is clean and the test
suite remains green.
