+++ kincir/examples/rabbitmq_ack_example.rs
@@ -1,51 +1,49 @@
-//! Kafka Acknowledgment Example
+//! RabbitMQ Acknowledgment Example
 //!
-//! This example demonstrates how to use Apache Kafka with manual acknowledgment handling.
+//! This example demonstrates how to use RabbitMQ with manual acknowledgment handling.
 //! It shows publishing messages, receiving them with acknowledgment handles, and
 //! performing manual acknowledgment or negative acknowledgment operations.
 //!
 //! Prerequisites:
-//! - Apache Kafka server running on localhost:9092
-//! - Topic creation (topics will be auto-created if auto.create.topics.enable=true)
+//! - RabbitMQ server running on localhost:5672
+//! - Default guest/guest credentials
 //!
-//! Run with: cargo run --example kafka_ack_example
+//! Run with: cargo run --example rabbitmq_ack_example
 
 use kincir::ack::{AckHandle, AckSubscriber};
-use kincir::kafka::{KafkaAckSubscriber, KafkaPublisher};
+use kincir::rabbitmq::{RabbitMQAckSubscriber, RabbitMQPublisher};
 use kincir::{Message, Publisher};
 use std::time::Duration;
 use tokio::time::{sleep, timeout};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    println!("🚀 Kafka Acknowledgment Example");
-    println!("=================================");
+    println!("🚀 RabbitMQ Acknowledgment Example");
+    println!("=====================================");
 
-    // Check if Kafka is available
-    if !is_kafka_available().await {
-        eprintln!("❌ Kafka is not available at localhost:9092");
-        eprintln!("   Please start Kafka server and try again.");
-        eprintln!("   Docker: docker run -d --name kafka -p 9092:9092 apache/kafka:latest");
+    // Check if RabbitMQ is available
+    if !is_rabbitmq_available().await {
+        eprintln!("❌ RabbitMQ is not available at localhost:5672");
+        eprintln!("   Please start RabbitMQ server and try again.");
+        eprintln!("   Docker: docker run -d --name rabbitmq -p 5672:5672 -p 15672:15672 rabbitmq:3-management");
         return Ok(());
     }
 
-    println!("✅ Kafka connection available");
+    println!("✅ RabbitMQ connection available");
 
     // Configuration
-    let brokers = vec!["127.0.0.1:9092".to_string()];
-    let topic = "ack-example-topic";
-    let consumer_group = "ack-example-group";
+    let rabbitmq_uri = "amqp://guest:guest@127.0.0.1:5672";
+    let topic = "ack-example-queue";
 
     // Create publisher
-    println!("\n📤 Creating Kafka publisher...");
-    let publisher = KafkaPublisher::new(brokers.clone())?;
+    println!("\n📤 Creating RabbitMQ publisher...");
+    let publisher = RabbitMQPublisher::new(rabbitmq_uri).await?;
     println!("✅ Publisher created successfully");
 
     // Create subscriber with acknowledgment support
-    println!("\n📥 Creating Kafka acknowledgment subscriber...");
-    let mut subscriber = KafkaAckSubscriber::new(brokers, consumer_group.to_string()).await?;
+    println!("\n📥 Creating RabbitMQ acknowledgment subscriber...");
+    let mut subscriber = RabbitMQAckSubscriber::new(rabbitmq_uri).await?;
     println!("✅ Acknowledgment subscriber created successfully");
-    println!("   Consumer Group: {}", subscriber.group_id());
 
     // Subscribe to topic
     println!("\n🔗 Subscribing to topic: {}", topic);
@@ -53,57 +51,57 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     println!("✅ Successfully subscribed to topic");
 
     // Example 1: Basic acknowledgment
-    println!("\n" + "=".repeat(50).as_str());
+    println!("\n{}", "=".repeat(50));
     println!("📋 Example 1: Basic Message Acknowledgment");
-    println!("=".repeat(50));
+    println!("{}", "=".repeat(50));
     
     basic_acknowledgment_example(&publisher, &mut subscriber, topic).await?;
 
     // Example 2: Negative acknowledgment with requeue
-    println!("\n" + "=".repeat(50).as_str());
+    println!("\n{}", "=".repeat(50));
     println!("📋 Example 2: Negative Acknowledgment with Requeue");
-    println!("=".repeat(50));
+    println!("{}", "=".repeat(50));
     
     negative_acknowledgment_example(&publisher, &mut subscriber, topic).await?;
 
     // Example 3: Batch acknowledgment
-    println!("\n" + "=".repeat(50).as_str());
+    println!("\n{}", "=".repeat(50));
     println!("📋 Example 3: Batch Acknowledgment");
-    println!("=".repeat(50));
+    println!("{}", "=".repeat(50));
     
     batch_acknowledgment_example(&publisher, &mut subscriber, topic).await?;
 
-    // Example 4: Offset management and consumer groups
-    println!("\n" + "=".repeat(50).as_str());
-    println!("📋 Example 4: Offset Management and Consumer Groups");
-    println!("=".repeat(50));
+    // Example 4: Error handling and retry logic
+    println!("\n{}", "=".repeat(50));
+    println!("📋 Example 4: Error Handling and Retry Logic");
+    println!("{}", "=".repeat(50));
     
-    offset_management_example(&publisher, &mut subscriber, topic).await?;
+    error_handling_example(&publisher, &mut subscriber, topic).await?;
 
     println!("\n🎉 All examples completed successfully!");
-    println!("   The Kafka acknowledgment system provides reliable message processing");
-    println!("   with manual control over offset commits and consumer group coordination.");
+    println!("   The RabbitMQ acknowledgment system provides reliable message processing");
+    println!("   with manual control over message acknowledgment and requeue behavior.");
 
     Ok(())
 }
 
 async fn basic_acknowledgment_example(
-    publisher: &KafkaPublisher,
-    subscriber: &mut KafkaAckSubscriber,
+    publisher: &RabbitMQPublisher,
+    subscriber: &mut RabbitMQAckSubscriber,
     topic: &str,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     println!("📤 Publishing message for basic acknowledgment...");
     
-    let message = Message::new(b"Hello from Kafka with acknowledgment!".to_vec())
+    let message = Message::new(b"Hello from RabbitMQ with acknowledgment!".to_vec())
         .with_metadata("example", "basic_ack")
-        .with_metadata("timestamp", &chrono::Utc::now().to_rfc3339());
+        .with_metadata("timestamp", chrono::Utc::now().to_rfc3339());
     
     publisher.publish(topic, vec![message.clone()]).await?;
     println!("✅ Message published: {}", String::from_utf8_lossy(&message.payload));
 
     println!("📥 Receiving message with acknowledgment handle...");
     let (received_message, ack_handle) = timeout(
-        Duration::from_secs(15),
+        Duration::from_secs(10),
         subscriber.receive_with_ack()
     ).await??;
 
@@ -115,22 +113,20 @@ async fn basic_acknowledgment_example(
     println!("📋 Acknowledgment handle details:");
     println!("   Message ID: {}", ack_handle.message_id());
     println!("   Topic: {}", ack_handle.topic());
-    println!("   Partition: {}", ack_handle.partition());
-    println!("   Offset: {}", ack_handle.offset());
     println!("   Delivery Count: {}", ack_handle.delivery_count());
     println!("   Is Retry: {}", ack_handle.is_retry());
     println!("   Handle ID: {}", ack_handle.handle_id());
 
-    println!("✅ Acknowledging message (committing offset)...");
+    println!("✅ Acknowledging message...");
     subscriber.ack(ack_handle).await?;
-    println!("✅ Message acknowledged successfully - offset committed");
+    println!("✅ Message acknowledged successfully");
 
     Ok(())
 }
 
 async fn negative_acknowledgment_example(
-    publisher: &KafkaPublisher,
-    subscriber: &mut KafkaAckSubscriber,
+    publisher: &RabbitMQPublisher,
+    subscriber: &mut RabbitMQAckSubscriber,
     topic: &str,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     println!("📤 Publishing message for negative acknowledgment...");
@@ -144,12 +140,11 @@ async fn negative_acknowledgment_example(
 
     println!("📥 Receiving message...");
     let (received_message, ack_handle) = timeout(
-        Duration::from_secs(15),
+        Duration::from_secs(10),
         subscriber.receive_with_ack()
     ).await??;
 
     println!("✅ Message received: {}", String::from_utf8_lossy(&received_message.payload));
-    println!("   Partition: {}, Offset: {}", ack_handle.partition(), ack_handle.offset());
 
     // Simulate processing failure
     let should_process = received_message.metadata.get("should_process")
@@ -159,9 +154,10 @@ async fn negative_acknowledgment_example(
     if !should_process {
         println!("❌ Processing failed - negatively acknowledging with requeue...");
         subscriber.nack(ack_handle, true).await?;
-        println!("✅ Message negatively acknowledged (offset NOT committed)");
-        println!("💡 Message will be redelivered on consumer restart or rebalance");
-        println!("   In Kafka, 'requeue' means we don't commit the offset");
+        println!("✅ Message negatively acknowledged and requeued");
+        
+        // The message is now back in the queue and could be received again
+        println!("💡 Message has been requeued and is available for reprocessing");
     } else {
         println!("✅ Processing successful - acknowledging...");
         subscriber.ack(ack_handle).await?;
@@ -171,8 +167,8 @@ async fn negative_acknowledgment_example(
 }
 
 async fn batch_acknowledgment_example(
-    publisher: &KafkaPublisher,
-    subscriber: &mut KafkaAckSubscriber,
+    publisher: &RabbitMQPublisher,
+    subscriber: &mut RabbitMQAckSubscriber,
     topic: &str,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     println!("📤 Publishing batch of messages...");
@@ -191,85 +187,86 @@ async fn batch_acknowledgment_example(
     
     for i in 0..messages.len() {
         let (received_message, ack_handle) = timeout(
-            Duration::from_secs(15),
+            Duration::from_secs(10),
             subscriber.receive_with_ack()
         ).await??;
         
-        println!("✅ Received message {}: {} (partition: {}, offset: {})", 
+        println!("✅ Received message {}: {}", 
                 i + 1, 
-                String::from_utf8_lossy(&received_message.payload),
-                ack_handle.partition(),
-                ack_handle.offset());
+                String::from_utf8_lossy(&received_message.payload));
         
         ack_handles.push(ack_handle);
     }
 
     println!("📋 Performing batch acknowledgment for {} messages...", ack_handles.len());
-    println!("   Kafka will commit the highest offset for each partition");
-    
     subscriber.ack_batch(ack_handles).await?;
     println!("✅ Batch acknowledgment completed successfully");
-    println!("💡 All messages in the batch were acknowledged with optimized offset commits");
+    println!("💡 All messages in the batch were acknowledged with a single operation");
 
     Ok(())
 }
 
-async fn offset_management_example(
-    publisher: &KafkaPublisher,
-    subscriber: &mut KafkaAckSubscriber,
+async fn error_handling_example(
+    publisher: &RabbitMQPublisher,
+    subscriber: &mut RabbitMQAckSubscriber,
     topic: &str,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    println!("📤 Publishing messages for offset management demonstration...");
+    println!("📤 Publishing message for error handling demonstration...");
     
-    let messages = vec![
-        Message::new(b"Offset message 1".to_vec()).with_metadata("process", "true"),
-        Message::new(b"Offset message 2".to_vec()).with_metadata("process", "false"),
-        Message::new(b"Offset message 3".to_vec()).with_metadata("process", "true"),
-    ];
+    let message = Message::new(b"Message for error handling".to_vec())
+        .with_metadata("simulate_error", "true")
+        .with_metadata("max_retries", "3");
     
-    publisher.publish(topic, messages.clone()).await?;
-    println!("✅ Published {} messages for offset management", messages.len());
+    publisher.publish(topic, vec![message.clone()]).await?;
+    println!("✅ Message published: {}", String::from_utf8_lossy(&message.payload));
 
-    println!("📥 Processing messages with selective acknowledgment...");
+    println!("📥 Receiving message and simulating processing with retries...");
     
-    for i in 0..messages.len() {
+    let mut retry_count = 0;
+    let max_retries = 3;
+    
+    loop {
         let (received_message, ack_handle) = timeout(
-            Duration::from_secs(15),
+            Duration::from_secs(10),
             subscriber.receive_with_ack()
         ).await??;
 
-        println!("🔄 Processing message {} at offset {}", i + 1, ack_handle.offset());
+        retry_count += 1;
+        println!("🔄 Processing attempt {} of {}", retry_count, max_retries);
         
-        // Simulate processing based on metadata
-        let should_process = received_message.metadata.get("process")
+        // Simulate processing
+        let simulate_error = received_message.metadata.get("simulate_error")
             .map(|v| v == "true")
             .unwrap_or(false);
 
-        if should_process {
-            println!("✅ Processing successful - acknowledging offset {}", ack_handle.offset());
-            subscriber.ack(ack_handle).await?;
-        } else {
-            println!("❌ Processing failed - discarding message at offset {}", ack_handle.offset());
-            // Discard by committing offset (skip this message)
+        if simulate_error && retry_count < max_retries {
+            println!("❌ Processing failed (simulated) - attempt {}", retry_count);
+            println!("🔄 Negatively acknowledging with requeue for retry...");
+            subscriber.nack(ack_handle, true).await?;
+            
+            // Add delay before retry
+            sleep(Duration::from_millis(500)).await;
+            continue;
+        } else if simulate_error && retry_count >= max_retries {
+            println!("❌ Max retries exceeded - negatively acknowledging without requeue");
+            println!("💀 Message will be discarded or sent to dead letter queue");
             subscriber.nack(ack_handle, false).await?;
+            break;
+        } else {
+            println!("✅ Processing successful - acknowledging message");
+            subscriber.ack(ack_handle).await?;
+            break;
         }
-        
-        // Small delay to demonstrate sequential processing
-        sleep(Duration::from_millis(100)).await;
     }
 
-    println!("📊 Offset management completed:");
-    println!("   Consumer group: {}", subscriber.group_id());
-    println!("   Topic: {}", topic);
-    println!("💡 Kafka tracks committed offsets per consumer group and partition");
-    println!("   This enables reliable message processing and consumer failover");
+    println!("📊 Error handling completed:");
+    println!("   Total attempts: {}", retry_count);
+    println!("   Max retries: {}", max_retries);
+    println!("💡 This demonstrates how to implement retry logic with acknowledgments");
 
     Ok(())
 }
 
-async fn is_kafka_available() -> bool {
-    match tokio::net::TcpStream::connect("127.0.0.1:9092").await {
-        Ok(_) => true,
-        Err(_) => false,
-    }
+async fn is_rabbitmq_available() -> bool {
+    (tokio::net::TcpStream::connect("127.0.0.1:5672").await).is_ok()
 }
